+++ mne/fiff/__init__.py
@@ -11,7 +11,8 @@ from .evoked import Evoked, read_evoked, write_evoked
 from .raw import Raw, read_raw_segment, read_raw_segment_times, \
                  start_writing_raw, write_raw_buffer, finish_writing_raw
 from .pick import pick_types, pick_channels, pick_types_evoked, \
-                  pick_channels_regexp
+                  pick_channels_regexp, pick_channels_forward, \
+                  pick_types_forward
 
 from .compensator import get_current_comp
 from .proj import compute_spatial_vectors